+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* (4 April 2016). Display Bulk User Import from CSV Form on Add User Page.
 * (30 November 2015). Merged Add new and existing user forms.
 * (16 June 2015). Cleaned up obsolete code.
 * (09 February 2015). Add existing user without confirmation.